+++ models/Document.php
@@ -18,10 +18,12 @@
 
 namespace Pimcore\Model;
 
+use Doctrine\DBAL\Exception\DeadlockException;
 use Pimcore\Event\DocumentEvents;
 use Pimcore\Event\FrontendEvents;
 use Pimcore\Event\Model\DocumentEvent;
 use Pimcore\Logger;
+use Pimcore\Model\Document\Hardlink;
 use Pimcore\Model\Document\Hardlink\Wrapper\WrapperInterface;
 use Pimcore\Model\Document\Listing;
 use Pimcore\Model\Element\ElementInterface;
@@ -33,6 +35,8 @@ use Symfony\Component\EventDispatcher\GenericEvent;
 /**
  * @method \Pimcore\Model\Document\Dao getDao()
  * @method bool __isBasedOnLatestData()
+ * @method int getChildAmount($user = null)
+ * @method string getCurrentFullPath()
  */
 class Document extends Element\AbstractElement
 {
@@ -49,7 +53,7 @@ class Document extends Element\AbstractElement
     private static $hideUnpublished = false;
 
     /**
-     * @var string
+     * @var string|null
      */
     protected $fullPathCache;
 
@@ -70,7 +74,7 @@ class Document extends Element\AbstractElement
     /**
      * The parent document.
      *
-     * @var Document
+     * @var Document|null
      */
     protected $parent;
 
@@ -139,23 +143,16 @@ class Document extends Element\AbstractElement
     protected $userModification;
 
     /**
-     * Permissions for the user which requested this document in editmode*
-     *
-     * @var array
-     */
-    protected $userPermissions;
-
-    /**
      * Dependencies for this document
      *
-     * @var Dependency
+     * @var Dependency|null
      */
     protected $dependencies;
 
     /**
      * List of Property, concerning the folder
      *
-     * @var array
+     * @var array|null
      */
     protected $properties = null;
 
@@ -188,9 +185,9 @@ class Document extends Element\AbstractElement
     protected $hasSiblings = [];
 
     /**
-     * Check if the document is locked.
+     * enum('self','propagate') nullable
      *
-     * @var string
+     * @var string|null
      */
     protected $locked = null;
 
@@ -213,7 +210,7 @@ class Document extends Element\AbstractElement
      * @param string $path
      * @param bool $force
      *
-     * @return Document|Document\Email|Document\Folder|Document\Hardlink|Document\Link|Document\Page|Document\Printcontainer|Document\Printpage|Document\Snippet
+     * @return static|null
      */
     public static function getByPath($path, $force = false)
     {
@@ -262,16 +259,16 @@ class Document extends Element\AbstractElement
      * @param int $id
      * @param bool $force
      *
-     * @return Document|Document\Email|Document\Folder|Document\Hardlink|Document\Link|Document\Page|Document\Printcontainer|Document\Printpage|Document\Snippet|Document\Newsletter|null
+     * @return static|null
      */
     public static function getById($id, $force = false)
     {
         if (!is_numeric($id) || $id < 1) {
             return null;
         }
-        $id = intval($id);
 
-        $cacheKey = 'document_' . $id;
+        $id = intval($id);
+        $cacheKey = self::getCacheKey($id);
 
         if (!$force && \Pimcore\Cache\Runtime::isRegistered($cacheKey)) {
             $document = \Pimcore\Cache\Runtime::get($cacheKey);
@@ -296,10 +293,14 @@ class Document extends Element\AbstractElement
                     }
                 }
 
+                /** @var Document $document */
                 $document = self::getModelFactory()->build($className);
                 \Pimcore\Cache\Runtime::set($cacheKey, $document);
+
                 $document->getDao()->getById($id);
                 $document->__setDataVersionTimestamp($document->getModificationDate());
+
+                $document->resetDirtyMap();
 
                 \Pimcore\Cache::save($document, $cacheKey);
             } else {
@@ -323,7 +324,7 @@ class Document extends Element\AbstractElement
      * @param array $data
      * @param bool $save
      *
-     * @return Document
+     * @return static
      */
     public static function create($parentId, $data = [], $save = true)
     {
@@ -353,8 +354,8 @@ class Document extends Element\AbstractElement
     public static function getList($config = [])
     {
         if (is_array($config)) {
-            $listClass = Listing::class;
-            $list = self::getModelFactory()->build($listClass);
+            /** @var Listing $list */
+            $list = self::getModelFactory()->build(Listing::class);
             $list->setValues($config);
 
             return $list;
@@ -449,7 +450,7 @@ class Document extends Element\AbstractElement
                     }
 
                     // we try to start the transaction $maxRetries times again (deadlocks, ...)
-                    if ($retries < ($maxRetries - 1)) {
+                    if ($e instanceof DeadlockException && $retries < ($maxRetries - 1)) {
                         $run = $retries + 1;
                         $waitTime = rand(1, 5) * 100000; // microseconds
                         Logger::warn('Unable to finish transaction (' . $run . ". run) because of the following reason '" . $e->getMessage() . "'. --> Retrying in " . $waitTime . ' microseconds ... (' . ($run + 1) . ' of ' . $maxRetries . ')');
@@ -599,7 +600,7 @@ class Document extends Element\AbstractElement
         $this->getDao()->update();
 
         //set document to registry
-        \Pimcore\Cache\Runtime::set('document_' . $this->getId(), $this);
+        \Pimcore\Cache\Runtime::set(self::getCacheKey($this->getId()), $this);
     }
 
     /**
@@ -621,7 +622,7 @@ class Document extends Element\AbstractElement
     public function clearDependentCache($additionalTags = [])
     {
         try {
-            $tags = ['document_' . $this->getId(), 'document_properties', 'output'];
+            $tags = [$this->getCacheTag(), 'document_properties', 'output'];
             $tags = array_merge($tags, $additionalTags);
 
             \Pimcore\Cache::clearTags($tags);
@@ -648,17 +649,18 @@ class Document extends Element\AbstractElement
      * set the children of the document
      *
      * @param self[] $children
+     * @param bool $includingUnpublished
      *
      * @return $this
      */
-    public function setChildren($children)
+    public function setChildren($children, $includingUnpublished = false)
     {
         if (empty($children)) {
             // unset all cached children
             $this->hasChildren = [];
             $this->children = [];
         } elseif (is_array($children)) {
-            $cacheKey = $this->getListingCacheKey();
+            $cacheKey = $this->getListingCacheKey([$includingUnpublished]);
             $this->children[$cacheKey] = $children;
             $this->hasChildren[$cacheKey] = (bool) count($children);
         }
@@ -752,9 +754,9 @@ class Document extends Element\AbstractElement
     }
 
     /**
-     * Returns true if the element is locked
+     * enum('self','propagate') nullable
      *
-     * @return string
+     * @return string|null
      */
     public function getLocked()
     {
@@ -766,9 +768,9 @@ class Document extends Element\AbstractElement
     }
 
     /**
-     * Mark the document as locked.
+     * enum('self','propagate') nullable
      *
-     * @param string $locked
+     * @param string|null $locked
      *
      * @return Document
      */
@@ -780,45 +782,66 @@ class Document extends Element\AbstractElement
     }
 
     /**
-     * @param bool $isNested
-     *
      * @throws \Exception
      */
-    public function delete(bool $isNested = false)
+    protected function doDelete()
+    {
+        // remove children
+        if ($this->hasChildren()) {
+            // delete also unpublished children
+            $unpublishedStatus = self::doHideUnpublished();
+            self::setHideUnpublished(false);
+            foreach ($this->getChildren(true) as $child) {
+                if (!$child instanceof WrapperInterface) {
+                    $child->delete();
+                }
+            }
+            self::setHideUnpublished($unpublishedStatus);
+        }
+
+        // remove all properties
+        $this->getDao()->deleteAllProperties();
+
+        // remove permissions
+        $this->getDao()->deleteAllPermissions();
+
+        // remove dependencies
+        $d = $this->getDependencies();
+        $d->cleanAllForElement($this);
+
+        // remove translations
+        $service = new Document\Service;
+        $service->removeTranslation($this);
+    }
+
+    /**
+     * @throws \Exception
+     */
+    public function delete()
     {
         \Pimcore::getEventDispatcher()->dispatch(DocumentEvents::PRE_DELETE, new DocumentEvent($this));
 
         $this->beginTransaction();
 
         try {
-            // remove children
-            if ($this->hasChildren()) {
-                // delete also unpublished children
-                $unpublishedStatus = self::doHideUnpublished();
-                self::setHideUnpublished(false);
-                foreach ($this->getChildren(true) as $child) {
-                    $child->delete(true);
-                }
-                self::setHideUnpublished($unpublishedStatus);
+            if ($this->getId() == 1) {
+                throw new \Exception('root-node cannot be deleted');
             }
 
-            // remove all properties
-            $this->getDao()->deleteAllProperties();
-
-            // remove permissions
-            $this->getDao()->deleteAllPermissions();
-
-            // remove dependencies
-            $d = $this->getDependencies();
-            $d->cleanAllForElement($this);
-
-            // remove translations
-            $service = new Document\Service;
-            $service->removeTranslation($this);
-
+            $this->doDelete();
             $this->getDao()->delete();
 
             $this->commit();
+
+            //clear parent data from registry
+            $parentCacheKey = self::getCacheKey($this->getParentId());
+            if (\Pimcore\Cache\Runtime::isRegistered($parentCacheKey)) {
+                /** @var Document $parent * */
+                $parent = \Pimcore\Cache\Runtime::get($parentCacheKey);
+                if ($parent instanceof self) {
+                    $parent->setChildren(null);
+                }
+            }
         } catch (\Exception $e) {
             $this->rollBack();
             $failureEvent = new DocumentEvent($this);
@@ -832,13 +855,13 @@ class Document extends Element\AbstractElement
         $this->clearDependentCache();
 
         //clear document from registry
-        \Pimcore\Cache\Runtime::set('document_' . $this->getId(), null);
+        \Pimcore\Cache\Runtime::set(self::getCacheKey($this->getId()), null);
 
         \Pimcore::getEventDispatcher()->dispatch(DocumentEvents::POST_DELETE, new DocumentEvent($this));
     }
 
     /**
-     * Returns the full path of the document including the key (path+key)
+     * Returns the frontend path to the document respecting the current site and pretty-URLs
      *
      * @param bool $force
      *
@@ -890,7 +913,7 @@ class Document extends Element\AbstractElement
             }
 
             if (!$link) {
-                $config = \Pimcore\Config::getSystemConfig();
+                $config = \Pimcore\Config::getSystemConfiguration('general');
                 $request = $requestStack->getCurrentRequest();
                 $scheme = 'http://';
                 if ($request) {
@@ -910,8 +933,8 @@ class Document extends Element\AbstractElement
                     }
                 }
 
-                if (!$link && $config->general->domain) {
-                    $link = $scheme . $config->general->domain . $this->getRealFullPath();
+                if (!$link && !empty($config['domain'])) {
+                    $link = $scheme . $config['domain'] . $this->getRealFullPath();
                 }
             }
         }
@@ -920,7 +943,7 @@ class Document extends Element\AbstractElement
             $link = $this->getPath() . $this->getKey();
         }
 
-        if($masterRequest) {
+        if ($masterRequest) {
             // caching should only be done when master request is available as it is done for performance reasons
             // of the web frontend, without a request object there's no need to cache anything
             // for details also see https://github.com/pimcore/pimcore/issues/5707
@@ -943,7 +966,7 @@ class Document extends Element\AbstractElement
             $path = urlencode_ignore_slash($path);
 
             $event = new GenericEvent($this, [
-                'frontendPath' => $path
+                'frontendPath' => $path,
             ]);
             \Pimcore::getEventDispatcher()->dispatch(FrontendEvents::DOCUMENT_PATH, $event);
             $path = $event->getArgument('frontendPath');
@@ -1042,7 +1065,7 @@ class Document extends Element\AbstractElement
     }
 
     /**
-     * Returns the full real path of the document.
+     * Returns the internal real full path of the document. (not for frontend use!)
      *
      * @return string
      */
@@ -1084,7 +1107,7 @@ class Document extends Element\AbstractElement
     /**
      * Set the document key.
      *
-     * @param int $key
+     * @param string $key
      *
      * @return Document
      */
@@ -1104,6 +1127,8 @@ class Document extends Element\AbstractElement
      */
     public function setModificationDate($modificationDate)
     {
+        $this->markFieldDirty('modificationDate');
+
         $this->modificationDate = (int) $modificationDate;
 
         return $this;
@@ -1177,7 +1202,7 @@ class Document extends Element\AbstractElement
     /**
      * Set the document type.
      *
-     * @param int $type
+     * @param string $type
      *
      * @return Document
      */
@@ -1217,6 +1242,8 @@ class Document extends Element\AbstractElement
      */
     public function setUserModification($userModification)
     {
+        $this->markFieldDirty('userModification');
+
         $this->userModification = (int) $userModification;
 
         return $this;
@@ -1372,7 +1399,7 @@ class Document extends Element\AbstractElement
     {
         $finalVars = [];
         $parentVars = parent::__sleep();
-        $blockedVars = ['dependencies', 'userPermissions', 'hasChildren', 'versions', 'scheduledTasks', 'parent', 'fullPathCache'];
+        $blockedVars = ['dependencies', 'hasChildren', 'versions', 'scheduledTasks', 'parent', 'fullPathCache'];
 
         if ($this->isInDumpState()) {
             // this is if we want to make a full dump of the object (eg. for a new version), including children for recyclebin
@@ -1434,7 +1461,7 @@ class Document extends Element\AbstractElement
         $this->removeInheritedProperties();
 
         // add to registry to avoid infinite regresses in the following $this->getDao()->getProperties()
-        $cacheKey = 'document_' . $this->getId();
+        $cacheKey = self::getCacheKey($this->getId());
         if (!\Pimcore\Cache\Runtime::isRegistered($cacheKey)) {
             \Pimcore\Cache\Runtime::set($cacheKey, $this);
         }
@@ -1477,14 +1504,6 @@ class Document extends Element\AbstractElement
     }
 
     /**
-     * @param array $userPermissions
-     */
-    public function setUserPermissions($userPermissions): void
-    {
-        $this->userPermissions = $userPermissions;
-    }
-
-    /**
      * @return int
      */
     public function getVersionCount(): int
@@ -1510,5 +1529,15 @@ class Document extends Element\AbstractElement
         $cacheKey = (string)$unpublished;
 
         return $cacheKey;
+    }
+
+    public function __clone()
+    {
+        parent::__clone();
+        $this->parent = null;
+        $this->hasSiblings = [];
+        $this->siblings = [];
+        $this->dependencies = null;
+        $this->fullPathCache = null;
     }
 }
